Replace vpicio_mts with new implementation
