+++ src/tests/vpicio_mts.c
@@ -32,6 +32,7 @@
 #include <math.h>
 #include <inttypes.h>
 #include "pdc.h"
+#include "pdc_timing.h"
 
 #define NPARTICLES 8388608
 
@@ -51,35 +52,26 @@ int
 main(int argc, char **argv)
 {
     int     rank = 0, size = 1;
-    pdcid_t pdc_id, cont_prop, cont_id;
-    pdcid_t obj_prop_xx, obj_prop_yy, obj_prop_zz, obj_prop_pxx, obj_prop_pyy, obj_prop_pzz, obj_prop_id11,
-        obj_prop_id22;
+    pdcid_t pdc_id, cont_prop, cont_id, region_local, region_remote;
+    pdcid_t obj_prop_float, obj_prop_int;
     pdcid_t obj_xx, obj_yy, obj_zz, obj_pxx, obj_pyy, obj_pzz, obj_id11, obj_id22;
-    pdcid_t region_x, region_y, region_z, region_px, region_py, region_pz, region_id1, region_id2;
-    pdcid_t region_xx, region_yy, region_zz, region_pxx, region_pyy, region_pzz, region_id11, region_id22;
-    perr_t  ret;
 #ifdef ENABLE_MPI
     MPI_Comm comm;
 #else
     int comm = 1;
 #endif
-    float *   x, *y, *z;
-    float *   px, *py, *pz;
-    int *     id1, *id2;
-    int       x_dim = 64;
-    int       y_dim = 64;
-    int       z_dim = 64;
-    uint64_t  numparticles, i;
-    uint64_t  dims[1];
-    int       ndim = 1;
-    uint64_t *offset;
-    uint64_t *offset_remote;
-    uint64_t *mysize;
-    double    t0, t1;
-    uint64_t  steps = 1, sleeptime = 0;
+    float *    x, *y, *z, *px, *py, *pz;
+    int *      id1, *id2;
+    int        x_dim = 64, y_dim = 64, z_dim = 64, ndim = 1, steps = 1, sleeptime = 0;
+    uint64_t   numparticles, dims[1], offset_local[1], offset_remote[1], mysize[1];
+    double     t0, t1;
+    char       cur_time[64];
+    time_t     t;
+    struct tm *tm;
 
     pdcid_t transfer_request_x, transfer_request_y, transfer_request_z, transfer_request_px,
         transfer_request_py, transfer_request_pz, transfer_request_id1, transfer_request_id2;
+    pdcid_t transfer_requests[8];
 
 #ifdef ENABLE_MPI
     MPI_Init(&argc, &argv);
@@ -95,15 +87,14 @@ main(int argc, char **argv)
         sleeptime    = atoi(argv[3]);
     }
     if (rank == 0)
-        printf("Writing %" PRIu64 " number of particles for %llu steps with %d clients.\n", numparticles,
-               steps, size);
+        printf("Writing %" PRIu64 " number of particles for %d steps with %d clients.\n", numparticles, steps,
+               size);
 
     dims[0] = numparticles * size;
 
-    x = (float *)malloc(numparticles * sizeof(float));
-    y = (float *)malloc(numparticles * sizeof(float));
-    z = (float *)malloc(numparticles * sizeof(float));
-
+    x  = (float *)malloc(numparticles * sizeof(float));
+    y  = (float *)malloc(numparticles * sizeof(float));
+    z  = (float *)malloc(numparticles * sizeof(float));
     px = (float *)malloc(numparticles * sizeof(float));
     py = (float *)malloc(numparticles * sizeof(float));
     pz = (float *)malloc(numparticles * sizeof(float));
@@ -118,45 +109,28 @@ main(int argc, char **argv)
     cont_prop = PDCprop_create(PDC_CONT_CREATE, pdc_id);
     if (cont_prop <= 0) {
         printf("Fail to create container property @ line  %d!\n", __LINE__);
-        return 1;
+        return FAIL;
     }
     // create a container
     cont_id = PDCcont_create_col("c1", cont_prop);
     if (cont_id <= 0) {
         printf("Fail to create container @ line  %d!\n", __LINE__);
-        return 1;
+        return FAIL;
     }
     // create an object property
-    obj_prop_xx = PDCprop_create(PDC_OBJ_CREATE, pdc_id);
+    obj_prop_float = PDCprop_create(PDC_OBJ_CREATE, pdc_id);
+    PDCprop_set_obj_dims(obj_prop_float, 1, dims);
+    PDCprop_set_obj_type(obj_prop_float, PDC_FLOAT);
+    PDCprop_set_obj_user_id(obj_prop_float, getuid());
+    PDCprop_set_obj_app_name(obj_prop_float, "VPICIO");
+    PDCprop_set_obj_tags(obj_prop_float, "tag0=1");
+    /* PDCprop_set_obj_transfer_region_type(obj_prop_float, PDC_REGION_LOCAL); */
+    PDCprop_set_obj_transfer_region_type(obj_prop_float, PDC_REGION_STATIC);
 
-    PDCprop_set_obj_dims(obj_prop_xx, 1, dims);
-    PDCprop_set_obj_type(obj_prop_xx, PDC_FLOAT);
-    PDCprop_set_obj_user_id(obj_prop_xx, getuid());
-    PDCprop_set_obj_app_name(obj_prop_xx, "VPICIO");
-    PDCprop_set_obj_tags(obj_prop_xx, "tag0=1");
+    obj_prop_int = PDCprop_obj_dup(obj_prop_float);
+    PDCprop_set_obj_type(obj_prop_int, PDC_INT);
 
-    obj_prop_yy = PDCprop_obj_dup(obj_prop_xx);
-    PDCprop_set_obj_type(obj_prop_yy, PDC_FLOAT);
-
-    obj_prop_zz = PDCprop_obj_dup(obj_prop_xx);
-    PDCprop_set_obj_type(obj_prop_zz, PDC_FLOAT);
-
-    obj_prop_pxx = PDCprop_obj_dup(obj_prop_xx);
-    PDCprop_set_obj_type(obj_prop_pxx, PDC_FLOAT);
-
-    obj_prop_pyy = PDCprop_obj_dup(obj_prop_xx);
-    PDCprop_set_obj_type(obj_prop_pyy, PDC_FLOAT);
-
-    obj_prop_pzz = PDCprop_obj_dup(obj_prop_xx);
-    PDCprop_set_obj_type(obj_prop_pzz, PDC_FLOAT);
-
-    obj_prop_id11 = PDCprop_obj_dup(obj_prop_xx);
-    PDCprop_set_obj_type(obj_prop_id11, PDC_INT);
-
-    obj_prop_id22 = PDCprop_obj_dup(obj_prop_xx);
-    PDCprop_set_obj_type(obj_prop_id22, PDC_INT);
-
-    for (i = 0; i < numparticles; i++) {
+    for (uint64_t i = 0; i < numparticles; i++) {
         id1[i] = i;
         id2[i] = i * 2;
         x[i]   = uniform_random_number() * x_dim;
@@ -167,463 +141,262 @@ main(int argc, char **argv)
         pz[i]  = ((float)id2[i] / numparticles) * z_dim;
     }
 
-    offset           = (uint64_t *)malloc(sizeof(uint64_t) * ndim);
-    offset_remote    = (uint64_t *)malloc(sizeof(uint64_t) * ndim);
-    mysize           = (uint64_t *)malloc(sizeof(uint64_t) * ndim);
-    offset[0]        = 0;
+    offset_local[0]  = 0;
     offset_remote[0] = rank * numparticles;
     mysize[0]        = numparticles;
 
-    // create a region
-    region_x   = PDCregion_create(ndim, offset, mysize);
-    region_y   = PDCregion_create(ndim, offset, mysize);
-    region_z   = PDCregion_create(ndim, offset, mysize);
-    region_px  = PDCregion_create(ndim, offset, mysize);
-    region_py  = PDCregion_create(ndim, offset, mysize);
-    region_pz  = PDCregion_create(ndim, offset, mysize);
-    region_id1 = PDCregion_create(ndim, offset, mysize);
-    region_id2 = PDCregion_create(ndim, offset, mysize);
+    // create local and remote region
+    region_local  = PDCregion_create(ndim, offset_local, mysize);
+    region_remote = PDCregion_create(ndim, offset_remote, mysize);
 
-    region_xx   = PDCregion_create(ndim, offset_remote, mysize);
-    region_yy   = PDCregion_create(ndim, offset_remote, mysize);
-    region_zz   = PDCregion_create(ndim, offset_remote, mysize);
-    region_pxx  = PDCregion_create(ndim, offset_remote, mysize);
-    region_pyy  = PDCregion_create(ndim, offset_remote, mysize);
-    region_pzz  = PDCregion_create(ndim, offset_remote, mysize);
-    region_id11 = PDCregion_create(ndim, offset_remote, mysize);
-    region_id22 = PDCregion_create(ndim, offset_remote, mysize);
-
-    for (i = 0; i < steps; i++) {
+    for (int iter = 0; iter < steps; iter++) {
 
 #ifdef ENABLE_MPI
         MPI_Barrier(MPI_COMM_WORLD);
+        PDC_get_time_str(cur_time);
+        if (rank == 0)
+            printf("\n[%s] #Step  %d\n", cur_time, iter);
         t0 = MPI_Wtime();
-        if (rank == 0) {
-            printf("\n#Step  %llu\n", i);
-        }
 #endif
-        PDCprop_set_obj_time_step(obj_prop_xx, i);
-        PDCprop_set_obj_time_step(obj_prop_yy, i);
-        PDCprop_set_obj_time_step(obj_prop_zz, i);
-        PDCprop_set_obj_time_step(obj_prop_pxx, i);
-        PDCprop_set_obj_time_step(obj_prop_pyy, i);
-        PDCprop_set_obj_time_step(obj_prop_pzz, i);
-        PDCprop_set_obj_time_step(obj_prop_id11, i);
-        PDCprop_set_obj_time_step(obj_prop_id22, i);
+        PDCprop_set_obj_time_step(obj_prop_float, iter);
+        PDCprop_set_obj_time_step(obj_prop_int, iter);
 
-        obj_xx = PDCobj_create_mpi(cont_id, "obj-var-xx", obj_prop_xx, 0, comm);
+        obj_xx = PDCobj_create_mpi(cont_id, "obj-var-xx", obj_prop_float, 0, comm);
         if (obj_xx == 0) {
-            printf("Error getting an object id of %s from server, exit...\n", "obj-var-xx");
-            exit(-1);
+            printf("Error getting an object id of %s from server\n", "x");
+            return FAIL;
         }
 
-        obj_yy = PDCobj_create_mpi(cont_id, "obj-var-yy", obj_prop_yy, 0, comm);
+        obj_yy = PDCobj_create_mpi(cont_id, "obj-var-yy", obj_prop_float, 0, comm);
         if (obj_yy == 0) {
-            printf("Error getting an object id of %s from server, exit...\n", "obj-var-yy");
-            exit(-1);
+            printf("Error getting an object id of %s from server\n", "y");
+            return FAIL;
         }
-        obj_zz = PDCobj_create_mpi(cont_id, "obj-var-zz", obj_prop_zz, 0, comm);
+        obj_zz = PDCobj_create_mpi(cont_id, "obj-var-zz", obj_prop_float, 0, comm);
         if (obj_zz == 0) {
-            printf("Error getting an object id of %s from server, exit...\n", "obj-var-zz");
-            exit(-1);
+            printf("Error getting an object id of %s from server\n", "z");
+            return FAIL;
         }
-        obj_pxx = PDCobj_create_mpi(cont_id, "obj-var-pxx", obj_prop_pxx, 0, comm);
+        obj_pxx = PDCobj_create_mpi(cont_id, "obj-var-pxx", obj_prop_float, 0, comm);
         if (obj_pxx == 0) {
-            printf("Error getting an object id of %s from server, exit...\n", "obj-var-pxx");
-            exit(-1);
+            printf("Error getting an object id of %s from server\n", "px");
+            return FAIL;
         }
-        obj_pyy = PDCobj_create_mpi(cont_id, "obj-var-pyy", obj_prop_pyy, 0, comm);
+        obj_pyy = PDCobj_create_mpi(cont_id, "obj-var-pyy", obj_prop_float, 0, comm);
         if (obj_pyy == 0) {
-            printf("Error getting an object id of %s from server, exit...\n", "obj-var-pyy");
-            exit(-1);
+            printf("Error getting an object id of %s from server\n", "py");
+            return FAIL;
         }
-        obj_pzz = PDCobj_create_mpi(cont_id, "obj-var-pzz", obj_prop_pzz, 0, comm);
+        obj_pzz = PDCobj_create_mpi(cont_id, "obj-var-pzz", obj_prop_float, 0, comm);
         if (obj_pzz == 0) {
-            printf("Error getting an object id of %s from server, exit...\n", "obj-var-pzz");
-            exit(-1);
+            printf("Error getting an object id of %s from server\n", "pz");
+            return FAIL;
         }
 
-        obj_id11 = PDCobj_create_mpi(cont_id, "id11", obj_prop_id11, 0, comm);
+        obj_id11 = PDCobj_create_mpi(cont_id, "id11", obj_prop_int, 0, comm);
         if (obj_id11 == 0) {
-            printf("Error getting an object id of %s from server, exit...\n", "obj_id11");
-            exit(-1);
+            printf("Error getting an object id of %s from server\n", "id1");
+            return FAIL;
         }
-        obj_id22 = PDCobj_create_mpi(cont_id, "id22", obj_prop_id22, 0, comm);
+        obj_id22 = PDCobj_create_mpi(cont_id, "id22", obj_prop_int, 0, comm);
         if (obj_id22 == 0) {
-            printf("Error getting an object id of %s from server, exit...\n", "obj_id22");
-            exit(-1);
+            printf("Error getting an object id of %s from server\n", "id2");
+            return FAIL;
         }
 
 #ifdef ENABLE_MPI
         MPI_Barrier(MPI_COMM_WORLD);
         t1 = MPI_Wtime();
-        if (rank == 0) {
-            printf("Obj create time: %.5e\n", t1 - t0);
-        }
+        PDC_get_time_str(cur_time);
+        if (rank == 0)
+            printf("[%s] Obj create time: %.5e\n", cur_time, t1 - t0);
 #endif
 
-        transfer_request_x = PDCregion_transfer_create(&x[0], PDC_WRITE, obj_xx, region_x, region_xx);
-        if (transfer_request_x == 0) {
-            printf("Array x transfer request creation failed\n");
-            return 1;
+        transfer_requests[0] =
+            PDCregion_transfer_create(&x[0], PDC_WRITE, obj_xx, region_local, region_remote);
+        if (transfer_requests[0] == 0) {
+            printf("x transfer request creation failed\n");
+            return FAIL;
         }
-        transfer_request_y = PDCregion_transfer_create(&y[0], PDC_WRITE, obj_yy, region_y, region_yy);
-        if (transfer_request_y == 0) {
-            printf("Array y transfer request creation failed\n");
-            return 1;
+        transfer_requests[1] =
+            PDCregion_transfer_create(&y[0], PDC_WRITE, obj_yy, region_local, region_remote);
+        if (transfer_requests[1] == 0) {
+            printf("y transfer request creation failed\n");
+            return FAIL;
         }
-        transfer_request_z = PDCregion_transfer_create(&z[0], PDC_WRITE, obj_zz, region_z, region_zz);
-        if (transfer_request_z == 0) {
-            printf("Array z transfer request creation failed\n");
-            return 1;
+        transfer_requests[2] =
+            PDCregion_transfer_create(&z[0], PDC_WRITE, obj_zz, region_local, region_remote);
+        if (transfer_requests[2] == 0) {
+            printf("z transfer request creation failed\n");
+            return FAIL;
         }
-        transfer_request_px = PDCregion_transfer_create(&px[0], PDC_WRITE, obj_pxx, region_px, region_pxx);
-        if (transfer_request_px == 0) {
-            printf("Array px transfer request creation failed\n");
-            return 1;
+        transfer_requests[3] =
+            PDCregion_transfer_create(&px[0], PDC_WRITE, obj_pxx, region_local, region_remote);
+        if (transfer_requests[3] == 0) {
+            printf("px transfer request creation failed\n");
+            return FAIL;
         }
-        transfer_request_py = PDCregion_transfer_create(&py[0], PDC_WRITE, obj_pyy, region_py, region_pyy);
-        if (transfer_request_py == 0) {
-            printf("Array py transfer request creation failed\n");
-            return 1;
+        transfer_requests[4] =
+            PDCregion_transfer_create(&py[0], PDC_WRITE, obj_pyy, region_local, region_remote);
+        if (transfer_requests[4] == 0) {
+            printf("py transfer request creation failed\n");
+            return FAIL;
         }
-        transfer_request_pz = PDCregion_transfer_create(&pz[0], PDC_WRITE, obj_pzz, region_pz, region_pzz);
-        if (transfer_request_pz == 0) {
-            printf("Array pz transfer request creation failed\n");
-            return 1;
+        transfer_requests[5] =
+            PDCregion_transfer_create(&pz[0], PDC_WRITE, obj_pzz, region_local, region_remote);
+        if (transfer_requests[5] == 0) {
+            printf("pz transfer request creation failed\n");
+            return FAIL;
         }
-        transfer_request_id1 =
-            PDCregion_transfer_create(&id1[0], PDC_WRITE, obj_id11, region_id1, region_id11);
-        if (transfer_request_id1 == 0) {
-            printf("Array id1 transfer request creation failed\n");
-            return 1;
+        transfer_requests[6] =
+            PDCregion_transfer_create(&id1[0], PDC_WRITE, obj_id11, region_local, region_remote);
+        if (transfer_requests[6] == 0) {
+            printf("id1 transfer request creation failed\n");
+            return FAIL;
         }
-        transfer_request_id2 =
-            PDCregion_transfer_create(&id2[0], PDC_WRITE, obj_id22, region_id2, region_id22);
-        if (transfer_request_id2 == 0) {
-            printf("Array id2 transfer request creation failed\n");
-            return 1;
+        transfer_requests[7] =
+            PDCregion_transfer_create(&id2[0], PDC_WRITE, obj_id22, region_local, region_remote);
+        if (transfer_requests[7] == 0) {
+            printf("id2 transfer request creation failed\n");
+            return FAIL;
         }
 
 #ifdef ENABLE_MPI
         MPI_Barrier(MPI_COMM_WORLD);
         t0 = MPI_Wtime();
-        if (rank == 0) {
-            printf("Transfer create time: %.5e\n", t0 - t1);
-        }
+        PDC_get_time_str(cur_time);
+        if (rank == 0)
+            printf("[%s] Transfer create time: %.5e\n", cur_time, t0 - t1);
 #endif
 
-        ret = PDCregion_transfer_start(transfer_request_x);
-        if (ret != SUCCEED) {
-            printf("Failed to start transfer for region_xx\n");
-            return 1;
-        }
-        ret = PDCregion_transfer_start(transfer_request_y);
-        if (ret != SUCCEED) {
-            printf("Failed to start transfer for region_yy\n");
-            return 1;
-        }
-        ret = PDCregion_transfer_start(transfer_request_z);
-        if (ret != SUCCEED) {
-            printf("Failed to start transfer for region_zz\n");
-            return 1;
-        }
-        ret = PDCregion_transfer_start(transfer_request_px);
-        if (ret != SUCCEED) {
-            printf("Failed to start transfer for region_pxx\n");
-            return 1;
-        }
-        ret = PDCregion_transfer_start(transfer_request_py);
-        if (ret != SUCCEED) {
-            printf("Failed to start transfer for region_pyy\n");
-            return 1;
-        }
-        ret = PDCregion_transfer_start(transfer_request_pz);
-        if (ret != SUCCEED) {
-            printf("Failed to start transfer for region_pzz\n");
-            return 1;
-        }
-        ret = PDCregion_transfer_start(transfer_request_id1);
-        if (ret != SUCCEED) {
-            printf("Failed to start transfer for region_id11\n");
-            return 1;
-        }
-        ret = PDCregion_transfer_start(transfer_request_id2);
-        if (ret != SUCCEED) {
-            printf("Failed to start transfer for region_id22\n");
-            return 1;
+        if (PDCregion_transfer_start_all(transfer_requests, 8) != SUCCEED) {
+            printf("Failed to start transfer requests\n");
+            return FAIL;
         }
 
 #ifdef ENABLE_MPI
         MPI_Barrier(MPI_COMM_WORLD);
         t1 = MPI_Wtime();
-        if (rank == 0) {
-            printf("Transfer start time: %.5e\n", t1 - t0);
-        }
+        PDC_get_time_str(cur_time);
+        if (rank == 0)
+            printf("[%s] Transfer start time: %.5e\n", cur_time, t1 - t0);
 #endif
-
-        ret = PDCregion_transfer_wait(transfer_request_x);
-        if (ret != SUCCEED) {
-            printf("Failed to transfer wait for region_xx\n");
-            return 1;
-        }
-        ret = PDCregion_transfer_wait(transfer_request_y);
-        if (ret != SUCCEED) {
-            printf("Failed to transfer wait for region_yy\n");
-            return 1;
-        }
-        ret = PDCregion_transfer_wait(transfer_request_z);
-        if (ret != SUCCEED) {
-            printf("Failed to transfer wait for region_zz\n");
-            return 1;
-        }
-        ret = PDCregion_transfer_wait(transfer_request_px);
-        if (ret != SUCCEED) {
-            printf("Failed to transfer wait for region_pxx\n");
-            return 1;
-        }
-        ret = PDCregion_transfer_wait(transfer_request_py);
-        if (ret != SUCCEED) {
-            printf("Failed to transfer wait for region_pyy\n");
-            return 1;
-        }
-        ret = PDCregion_transfer_wait(transfer_request_pz);
-        if (ret != SUCCEED) {
-            printf("Failed to transfer wait for region_pzz\n");
-            return 1;
-        }
-        ret = PDCregion_transfer_wait(transfer_request_id1);
-        if (ret != SUCCEED) {
-            printf("Failed to transfer wait for region_id11\n");
-            return 1;
-        }
-        ret = PDCregion_transfer_wait(transfer_request_id2);
-        if (ret != SUCCEED) {
-            printf("Failed to transfer wait for region_id22\n");
-            return 1;
-        }
-
-#ifdef ENABLE_MPI
-        MPI_Barrier(MPI_COMM_WORLD);
-        t0 = MPI_Wtime();
-        if (rank == 0) {
-            printf("Transfer wait time: %.5e\n", t0 - t1);
-        }
-#endif
-
-        ret = PDCregion_transfer_close(transfer_request_x);
-        if (ret != SUCCEED) {
-            printf("region xx transfer close failed\n");
-            return 1;
-        }
-        ret = PDCregion_transfer_close(transfer_request_y);
-        if (ret != SUCCEED) {
-            printf("region yy transfer close failed\n");
-            return 1;
-        }
-        ret = PDCregion_transfer_close(transfer_request_z);
-        if (ret != SUCCEED) {
-            printf("region zz transfer close failed\n");
-            return 1;
-        }
-        ret = PDCregion_transfer_close(transfer_request_px);
-        if (ret != SUCCEED) {
-            printf("region pxx transfer close failed\n");
-            return 1;
-        }
-        ret = PDCregion_transfer_close(transfer_request_py);
-        if (ret != SUCCEED) {
-            printf("region pyy transfer close failed\n");
-            return 1;
-        }
-        ret = PDCregion_transfer_close(transfer_request_pz);
-        if (ret != SUCCEED) {
-            printf("region pzz transfer close failed\n");
-            return 1;
-        }
-        ret = PDCregion_transfer_close(transfer_request_id1);
-        if (ret != SUCCEED) {
-            printf("region id11 transfer close failed\n");
-            return 1;
-        }
-        ret = PDCregion_transfer_close(transfer_request_id2);
-        if (ret != SUCCEED) {
-            printf("region id22 transfer close failed\n");
-            return 1;
-        }
-
-#ifdef ENABLE_MPI
-        MPI_Barrier(MPI_COMM_WORLD);
-        t1 = MPI_Wtime();
-        if (rank == 0) {
-            printf("Transfer close time: %.5e\n", t1 - t0);
-        }
-#endif
-
-        if (PDCobj_close(obj_xx) < 0) {
-            printf("fail to close obj_xx\n");
-            return 1;
-        }
-
-        if (PDCobj_close(obj_yy) < 0) {
-            printf("fail to close object obj_yy\n");
-            return 1;
-        }
-        if (PDCobj_close(obj_zz) < 0) {
-            printf("fail to close object obj_zz\n");
-            return 1;
-        }
-        if (PDCobj_close(obj_pxx) < 0) {
-            printf("fail to close object obj_pxx\n");
-            return 1;
-        }
-        if (PDCobj_close(obj_pyy) < 0) {
-            printf("fail to close object obj_pyy\n");
-            return 1;
-        }
-        if (PDCobj_close(obj_pzz) < 0) {
-            printf("fail to close object obj_pzz\n");
-            return 1;
-        }
-        if (PDCobj_close(obj_id11) < 0) {
-            printf("fail to close object obj_id11\n");
-            return 1;
-        }
-        if (PDCobj_close(obj_id22) < 0) {
-            printf("fail to close object obj_id22\n");
-            return 1;
-        }
-
-#ifdef ENABLE_MPI
-        MPI_Barrier(MPI_COMM_WORLD);
-        t0 = MPI_Wtime();
-        if (rank == 0) {
-            printf("Obj close time: %.5e\n", t0 - t1);
-        }
-#endif
-        if (i != steps - 1) {
+        // Emulate compute with sleep
+        if (iter != steps - 1) {
+            PDC_get_time_str(cur_time);
+            if (rank == 0)
+                printf("[%s] Sleep start: %llu.00\n", cur_time, sleeptime);
             sleep(sleeptime);
-            if (rank == 0) {
-                printf("Sleep time: %llu.00\n", sleeptime);
+            PDC_get_time_str(cur_time);
+            if (rank == 0)
+                printf("[%s] Sleep end: %llu.00\n", cur_time, sleeptime);
+        }
+
+#ifdef ENABLE_MPI
+        MPI_Barrier(MPI_COMM_WORLD);
+        t0 = MPI_Wtime();
+#endif
+
+        if (PDCregion_transfer_wait_all(transfer_requests, 8) != SUCCEED) {
+            printf("Failed to transfer wait all\n");
+            return FAIL;
+        }
+
+#ifdef ENABLE_MPI
+        MPI_Barrier(MPI_COMM_WORLD);
+        t1 = MPI_Wtime();
+        PDC_get_time_str(cur_time);
+        if (rank == 0)
+            printf("[%s] Transfer wait time: %.5e\n", cur_time, t1 - t0);
+#endif
+
+        for (int j = 0; j < 8; j++) {
+            if (PDCregion_transfer_close(transfer_requests[j]) != SUCCEED) {
+                printf("region transfer close failed\n");
+                return FAIL;
             }
         }
+
+#ifdef ENABLE_MPI
+        MPI_Barrier(MPI_COMM_WORLD);
+        t0 = MPI_Wtime();
+        PDC_get_time_str(cur_time);
+        if (rank == 0)
+            printf("[%s] Transfer close time: %.5e\n", cur_time, t0 - t1);
+#endif
+
+        if (PDCobj_close(obj_xx) != SUCCEED) {
+            printf("fail to close obj_xx\n");
+            return FAIL;
+        }
+        if (PDCobj_close(obj_yy) != SUCCEED) {
+            printf("fail to close object obj_yy\n");
+            return FAIL;
+        }
+        if (PDCobj_close(obj_zz) != SUCCEED) {
+            printf("fail to close object obj_zz\n");
+            return FAIL;
+        }
+        if (PDCobj_close(obj_pxx) != SUCCEED) {
+            printf("fail to close object obj_pxx\n");
+            return FAIL;
+        }
+        if (PDCobj_close(obj_pyy) != SUCCEED) {
+            printf("fail to close object obj_pyy\n");
+            return FAIL;
+        }
+        if (PDCobj_close(obj_pzz) != SUCCEED) {
+            printf("fail to close object obj_pzz\n");
+            return FAIL;
+        }
+        if (PDCobj_close(obj_id11) != SUCCEED) {
+            printf("fail to close object obj_id11\n");
+            return FAIL;
+        }
+        if (PDCobj_close(obj_id22) != SUCCEED) {
+            printf("fail to close object obj_id22\n");
+            return FAIL;
+        }
+
+#ifdef ENABLE_MPI
+        MPI_Barrier(MPI_COMM_WORLD);
+        t1 = MPI_Wtime();
+        PDC_get_time_str(cur_time);
+        if (rank == 0)
+            printf("[%s] Obj close time: %.5e\n", cur_time, t1 - t0);
+#endif
     } // End for steps
 
     PDC_timing_report("write");
 
-    if (PDCprop_close(obj_prop_xx) < 0) {
-        printf("Fail to close obj property obj_prop_xx\n");
-        return 1;
+    if (PDCprop_close(obj_prop_float) != SUCCEED) {
+        printf("Fail to close obj_prop_float\n");
+        return FAIL;
     }
-    if (PDCprop_close(obj_prop_yy) < 0) {
-        printf("Fail to close obj property obj_prop_yy\n");
-        return 1;
+    if (PDCprop_close(obj_prop_int) != SUCCEED) {
+        printf("Fail to close obj_prop_int\n");
+        return FAIL;
     }
-    if (PDCprop_close(obj_prop_zz) < 0) {
-        printf("Fail to close obj property obj_prop_zz\n");
-        return 1;
+    if (PDCregion_close(region_local) != SUCCEED) {
+        printf("fail to close local region \n");
+        return FAIL;
     }
-    if (PDCprop_close(obj_prop_pxx) < 0) {
-        printf("Fail to close obj property obj_prop_pxx\n");
-        return 1;
+    if (PDCobj_close(region_remote) != SUCCEED) {
+        printf("fail to close remote region\n");
+        return FAIL;
     }
-    if (PDCprop_close(obj_prop_pyy) < 0) {
-        printf("Fail to close obj property obj_prop_pyy\n");
-        return 1;
+    if (PDCcont_close(cont_id) != SUCCEED) {
+        printf("fail to close container\n");
+        return FAIL;
     }
-    if (PDCprop_close(obj_prop_pzz) < 0) {
-        printf("Fail to close obj property obj_prop_pzz\n");
-        return 1;
+    if (PDCprop_close(cont_prop) != SUCCEED) {
+        printf("Fail to close property\n");
+        return FAIL;
     }
-    if (PDCprop_close(obj_prop_id11) < 0) {
-        printf("Fail to close obj property obj_prop_id11\n");
-        return 1;
-    }
-    if (PDCprop_close(obj_prop_id22) < 0) {
-        printf("Fail to close obj property obj_prop_id22\n");
-        return 1;
-    }
-    if (PDCregion_close(region_x) < 0) {
-        printf("fail to close region region_x\n");
-        return 1;
-    }
-    if (PDCregion_close(region_y) < 0) {
-        printf("fail to close region region_y\n");
-        return 1;
-    }
-    if (PDCregion_close(region_z) < 0) {
-        printf("fail to close region region_z\n");
-        return 1;
-    }
-    if (PDCregion_close(region_px) < 0) {
-        printf("fail to close region region_px\n");
-        return 1;
-    }
-    if (PDCregion_close(region_py) < 0) {
-        printf("fail to close region region_py\n");
-        return 1;
-    }
-    if (PDCobj_close(region_pz) < 0) {
-        printf("fail to close region region_pz\n");
-        return 1;
-    }
-    if (PDCobj_close(region_id1) < 0) {
-        printf("fail to close region region_id1\n");
-        return 1;
-    }
-    if (PDCobj_close(region_id2) < 0) {
-        printf("fail to close region region_id2\n");
-        return 1;
-    }
-    if (PDCregion_close(region_xx) < 0) {
-        printf("fail to close region region_xx\n");
-        return 1;
-    }
-    if (PDCregion_close(region_yy) < 0) {
-        printf("fail to close region region_yy\n");
-        return 1;
-    }
-    if (PDCregion_close(region_zz) < 0) {
-        printf("fail to close region region_zz\n");
-        return 1;
-    }
-    if (PDCregion_close(region_pxx) < 0) {
-        printf("fail to close region region_pxx\n");
-        return 1;
-    }
-    if (PDCregion_close(region_pyy) < 0) {
-        printf("fail to close region region_pyy\n");
-        return 1;
-    }
-    if (PDCregion_close(region_pzz) < 0) {
-        printf("fail to close region region_pzz\n");
-        return 1;
-    }
-    if (PDCobj_close(region_id11) < 0) {
-        printf("fail to close region region_id11\n");
-        return 1;
-    }
-    if (PDCobj_close(region_id22) < 0) {
-        printf("fail to close region region_id22\n");
-        return 1;
-    }
-    // close a container
-    if (PDCcont_close(cont_id) < 0) {
-        printf("fail to close container c1\n");
-        return 1;
-    }
-    // close a container property
-    if (PDCprop_close(cont_prop) < 0) {
-        printf("Fail to close property @ line %d\n", __LINE__);
-        return 1;
-    }
-    if (PDCclose(pdc_id) < 0) {
+    if (PDCclose(pdc_id) != SUCCEED) {
         printf("fail to close PDC\n");
-        return 1;
+        return FAIL;
     }
-    free(offset);
-    free(offset_remote);
-    free(mysize);
     free(x);
     free(y);
     free(z);
